CAMEL-16643: Dropped the useless dependency to camel-xslt-saxon from camel-saxon
