+++ components/camel-saxon/pom.xml
@@ -46,11 +46,6 @@
             <artifactId>camel-xpath</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-xslt-saxon</artifactId>
-        </dependency>
-
-        <dependency>
             <groupId>net.sf.saxon</groupId>
             <artifactId>Saxon-HE</artifactId>
         </dependency>
