+++ components/camel-xslt-saxon/src/test/data/terminate.xml
