+++ components/camel-xslt-saxon/src/test/java/org/apache/camel/component/xslt/SaxonXsltDTDTest.java
@@ -88,10 +88,10 @@ public class SaxonXsltDTDTest extends CamelTestSupport {
     }
 
     @Override
-    protected RouteBuilder createRouteBuilder() throws Exception {
+    protected RouteBuilder createRouteBuilder() {
         return new RouteBuilder() {
             @Override
-            public void configure() throws Exception {
+            public void configure() {
 
                 from("direct:start1")
                         .to("xslt-saxon:org/apache/camel/component/xslt/transform_dtd.xsl")
