+++ components/camel-xslt-saxon/src/test/java/org/apache/camel/component/xslt/SaxonXsltFeatureRouteTest.java
@@ -28,7 +28,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 public class SaxonXsltFeatureRouteTest extends CamelTestSupport {
 
     @Test
-    public void testSendMessage() throws Exception {
+    public void testSendMessage() {
         String message = "<hello/>";
         sendXmlMessage("direct:start1", message);
         sendXmlMessage("direct:start2", message);
@@ -47,10 +47,10 @@ public class SaxonXsltFeatureRouteTest extends CamelTestSupport {
     }
 
     @Override
-    protected RouteBuilder createRouteBuilder() throws Exception {
+    protected RouteBuilder createRouteBuilder() {
         return new RouteBuilder() {
             @Override
-            public void configure() throws Exception {
+            public void configure() {
                 from("direct:start1")
                         .to("xslt-saxon:org/apache/camel/component/xslt/transform_text_imported.xsl")
                         .to("mock:result");
