+++ components/camel-xslt-saxon/src/test/java/org/apache/camel/component/xslt/SaxonXsltMessageTerminateTest.java
@@ -49,10 +49,10 @@ public class SaxonXsltMessageTerminateTest extends CamelTestSupport {
     }
 
     @Override
-    protected RouteBuilder createRouteBuilder() throws Exception {
+    protected RouteBuilder createRouteBuilder() {
         return new RouteBuilder() {
             @Override
-            public void configure() throws Exception {
+            public void configure() {
                 errorHandler(deadLetterChannel("mock:dead"));
 
                 from("file:src/test/data/?fileName=terminate.xml&noop=true").routeId("foo").noAutoStartup()
