+++ components/camel-xslt-saxon/src/test/java/org/apache/camel/component/xslt/extensions/MyExtensionFunction2.java
