+++ components/camel-xslt-saxon/src/test/java/org/apache/camel/component/xslt/extensions/SaxonExtensionFunctionsTest.java
