--- components/camel-xslt-saxon/src/test/java/org/apache/camel/component/xslt/saxon/MyXPath.java
@@ -1,42 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.camel.component.xslt.saxon;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.w3c.dom.NodeList;
-
-import org.apache.camel.language.xpath.XPathAnnotationExpressionFactory;
-import org.apache.camel.support.language.LanguageAnnotation;
-import org.apache.camel.support.language.NamespacePrefix;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER })
-@LanguageAnnotation(language = "xpath", factory = XPathAnnotationExpressionFactory.class)
-public @interface MyXPath {
-    String value();
-
-    // You can add the namespaces as the default value of the annotation
-    NamespacePrefix[] namespaces() default {
-            @NamespacePrefix(prefix = "n1", uri = "http://example.org/ns1"),
-            @NamespacePrefix(prefix = "n2", uri = "http://example.org/ns2") };
-
-    Class<?> resultType() default NodeList.class;
-}
