+++ components/camel-xslt-saxon/src/test/resources/org/apache/camel/component/xslt/transform_text.xsl
