+++ components/camel-xslt-saxon/src/test/resources/org/apache/camel/util/toolbox/aggregate.xsl
